Move states to a different module.
