+++ crates/bounce/src/states/mod.rs
@@ -1,0 +1,9 @@
+//! a module that contains different states that are supported by bounce.
+
+pub(crate) mod artifact;
+pub(crate) mod atom;
+pub(crate) mod future_notion;
+pub(crate) mod input_selector;
+pub(crate) mod notion;
+pub(crate) mod selector;
+pub(crate) mod slice;
